+++ R/utils.R
@@ -61,5 +61,6 @@ set_metadata <- function(x, ...) {
   else x
 }
 
-globalVariables(c(".", "Date", "type", "region", "key","type","value", "n",
-                  "...1"))
+globalVariables(
+  c(".", "Date", "type", "region", "key","type","value", "n","...1")
+)