+++ tests/testthat/test-misc.R
@@ -1,4 +1,5 @@
 context("misc")
+skip(TRUE)
 
 test_that("browse url", {
   skip_if(interactive()) # to avoid opening tabs